bzr revid: bde@tinyerp.com-20140212100909-rbfkncfa5i9gn7rv

+++ openerp/addons/base/ir/ir_model.py
@@ -3,7 +3,7 @@
 ##############################################################################
 #
 #    OpenERP, Open Source Business Applications
-#    Copyright (C) 2004-2012 OpenERP S.A. (<http://openerp.com>).
+#    Copyright (C) 2004-2014 OpenERP S.A. (<http://openerp.com>).
 #
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU Affero General Public License as
@@ -29,7 +29,7 @@ import openerp.modules.registry
 from openerp import SUPERUSER_ID
 from openerp import tools
 from openerp.osv import fields,osv
-from openerp.osv.orm import Model
+from openerp.osv.orm import Model, browse_null
 from openerp.tools.safe_eval import safe_eval as eval
 from openerp.tools import config
 from openerp.tools.translate import _
@@ -737,7 +737,7 @@ class ir_model_access(osv.osv):
                 msg_params = (model_name,)
             _logger.warning('Access Denied by ACLs for operation: %s, uid: %s, model: %s', mode, uid, model_name)
             msg = '%s %s' % (msg_heads[mode], msg_tail)
-            raise except_orm(_('Access Denied'), msg % msg_params)
+            raise openerp.exceptions.AccessError(msg % msg_params)
         return r or False
 
     __cache_clearing_methods = []
@@ -853,24 +853,58 @@ class ir_model_data(osv.osv):
         if not cr.fetchone():
             cr.execute('CREATE INDEX ir_model_data_module_name_index ON ir_model_data (module, name)')
 
-    @tools.ormcache()
+    # NEW V8 API
+    @tools.ormcache(skiparg=3)
+    def xmlid_lookup(self, cr, uid, xmlid):
+        """Low level xmlid lookup
+        Return (id, res_model, res_id) or raise ValueError if not found
+        """
+        module, name = xmlid.split('.', 1)
+        ids = self.search(cr, uid, [('module','=',module), ('name','=', name)])
+        if not ids:
+            raise ValueError('External ID not found in the system: %s' % (xmlid))
+        # the sql constraints ensure us we have only one result
+        res = self.read(cr, uid, ids[0], ['model', 'res_id'])
+        if not res['res_id']:
+            raise ValueError('External ID not found in the system: %s' % (xmlid))
+        return ids[0], res['model'], res['res_id']
+    
+    def xmlid_to_res_model_res_id(self, cr, uid, xmlid, raise_if_not_found=False):
+        """ Return (res_model, res_id)"""
+        try:
+            return self.xmlid_lookup(cr, uid, xmlid)[1:3]
+        except ValueError:
+            if raise_if_not_found:
+                raise
+            return (False, False)
+
+    def xmlid_to_res_id(self, cr, uid, xmlid, raise_if_not_found=False):
+        """ Returns res_id """
+        return self.xmlid_to_res_model_res_id(cr, uid, xmlid, raise_if_not_found)[1]
+
+    def xmlid_to_object(self, cr, uid, xmlid, raise_if_not_found=False, context=None):
+        """ Return a browse_record
+        if not found and raise_if_not_found is True return the browse_null
+        """ 
+        t = self.xmlid_to_res_model_res_id(cr, uid, xmlid, raise_if_not_found)
+        res_model, res_id = t
+
+        if res_model and res_id:
+            record = self.pool[res_model].browse(cr, uid, res_id, context=context)
+            if record.exists():
+                return record
+            if raise_if_not_found:
+                raise ValueError('No record found for unique ID %s. It may have been deleted.' % (xml_id))
+        return browse_null()
+
+    # OLD API
     def _get_id(self, cr, uid, module, xml_id):
         """Returns the id of the ir.model.data record corresponding to a given module and xml_id (cached) or raise a ValueError if not found"""
-        ids = self.search(cr, uid, [('module','=',module), ('name','=', xml_id)])
-        if not ids:
-            raise ValueError('No such external ID currently defined in the system: %s.%s' % (module, xml_id))
-        # the sql constraints ensure us we have only one result
-        return ids[0]
+        return self.xmlid_lookup(cr, uid, "%s.%s" % (module, xml_id))[0]
 
-    @tools.ormcache()
     def get_object_reference(self, cr, uid, module, xml_id):
         """Returns (model, res_id) corresponding to a given module and xml_id (cached) or raise ValueError if not found"""
-        data_id = self._get_id(cr, uid, module, xml_id)
-        #assuming data_id is not False, as it was checked upstream
-        res = self.read(cr, uid, data_id, ['model', 'res_id'])
-        if not res['res_id']:
-            raise ValueError('No such external ID currently defined in the system: %s.%s' % (module, xml_id))
-        return res['model'], res['res_id']
+        return self.xmlid_lookup(cr, uid, "%s.%s" % (module, xml_id))[1:3]
 
     def check_object_reference(self, cr, uid, module, xml_id, raise_on_access_error=False):
         """Returns (model, res_id) corresponding to a given module and xml_id (cached), if and only if the user has the necessary access rights
@@ -885,12 +919,11 @@ class ir_model_data(osv.osv):
         return model, False
 
     def get_object(self, cr, uid, module, xml_id, context=None):
-        """Returns a browsable record for the given module name and xml_id or raise ValueError if not found"""
-        res_model, res_id = self.get_object_reference(cr, uid, module, xml_id)
-        result = self.pool[res_model].browse(cr, uid, res_id, context=context)
-        if not result.exists():
-            raise ValueError('No record found for unique ID %s.%s. It may have been deleted.' % (module, xml_id))
-        return result
+        """ Returns a browsable record for the given module name and xml_id.
+            If not found, raise a ValueError or return a browse_null, depending
+            on the value of `raise_exception`.
+        """
+        return self.xmlid_to_object(cr, uid, "%s.%s" % (module, xml_id), raise_if_not_found=True, context=context)
 
     def _update_dummy(self,cr, uid, model, module, xml_id=False, store=True):
         if not xml_id:
@@ -907,8 +940,7 @@ class ir_model_data(osv.osv):
 
         :returns: itself
         """
-        self._get_id.clear_cache(self)
-        self.get_object_reference.clear_cache(self)
+        self.xmlid_lookup.clear_cache(self)
         return self
 
     def unlink(self, cr, uid, ids, context=None):
@@ -929,15 +961,17 @@ class ir_model_data(osv.osv):
             return False
         action_id = False
         if xml_id:
-            cr.execute('''SELECT imd.id, imd.res_id, md.id, imd.model
+            cr.execute('''SELECT imd.id, imd.res_id, md.id, imd.model, imd.noupdate
                           FROM ir_model_data imd LEFT JOIN %s md ON (imd.res_id = md.id)
                           WHERE imd.module=%%s AND imd.name=%%s''' % model_obj._table,
                           (module, xml_id))
             results = cr.fetchall()
-            for imd_id2,res_id2,real_id2,real_model in results:
+            for imd_id2,res_id2,real_id2,real_model,noupdate_imd in results:
+                # In update mode, do not update a record if it's ir.model.data is flagged as noupdate
+                if mode == 'update' and noupdate_imd:
+                    return res_id2
                 if not real_id2:
-                    self._get_id.clear_cache(self, uid, module, xml_id)
-                    self.get_object_reference.clear_cache(self, uid, module, xml_id)
+                    self.clear_caches()
                     cr.execute('delete from ir_model_data where id=%s', (imd_id2,))
                     res_id = False
                 else:
